competitor-api/.aws-sam/auto-dependency-layer/HelloWorldFunction/requirements.txt
snscrape==0.4.3.20220106
pandas==1.1.3
numpy==1.18.5
boto3==1.21.44
tweet-preprocessor==0.6.0
